Add: Required password when the account is deleted

+++ lib/utils/logout.dart
@@ -1,8 +1,34 @@
+import 'package:flutter/material.dart';
+import 'package:flutter_background_service/flutter_background_service.dart';
+import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:flutter_secure_storage/flutter_secure_storage.dart';
+import 'package:madnolia/blocs/chats/chats_bloc.dart';
+import 'package:madnolia/blocs/message/message_bloc.dart';
+import 'package:madnolia/blocs/player_matches/player_matches_bloc.dart';
+import 'package:madnolia/blocs/user/user_bloc.dart';
+import 'package:madnolia/database/providers/friendship_db.dart';
+import 'package:madnolia/database/providers/match_db.dart';
+import 'package:madnolia/database/providers/user_db.dart';
+import 'package:madnolia/services/sockets_service.dart';
 
-logoutApp() async {
+logoutApp(BuildContext context) async {
   const storage = FlutterSecureStorage();
+  final chatsBloc = context.read<ChatsBloc>();
+  final userBloc = context.read<UserBloc>();
+  final messageBloc = context.read<MessageBloc>();
+  final backgroundService = FlutterBackgroundService();
+  final matchesBloc = context.read<PlayerMatchesBloc>();
+  backgroundService.invoke('logout');
+  backgroundService.invoke("delete_all_notifications");
+  userBloc.logOutUser();
+  messageBloc.add(RestoreState());
+  matchesBloc.add(RestoreMatchesState());
+  chatsBloc.add(RestoreUserChats());
+  stopBackgroundService();
+  if(!context.mounted) return;
+  UserProvider.clearTable();
+  FriendshipProvider.deleteAll();
+  MatchProvider.deleteAll();
   await storage.deleteAll();
-
   
 }
